refactor: update test fixtures and enhance summaries integration
Test Updates:
- Update integration test configuration for summaries
- Refresh 152 OCR chunk sample JSON files with corrected data
- Add backup of test data (소장_p5-7.json.backup)
- Update sample data for 부동산소유권등기소송 case chunks

App Enhancements:
- Improve embedding generation for summaries
- Enhance Qdrant service for multi-collection support
- Update search functionality for summaries
- Refine Korean and multilingual sparse embedding
- Add configuration options for dual collection usage

These changes support the expanded summaries feature alongside
the existing chunks functionality, enabling semantic search
across both document chunks and generated summaries.

+++ app/korean_sparse_embedding.py
@@ -113,7 +113,8 @@ class KoreanSparseEmbedding:
             values = list(sparse_dict.values())
             norm = np.linalg.norm(values)
             if norm > 0:
-                sparse_dict = {k: v / norm for k, v in sparse_dict.items()}
+                # Ensure Python float type (not numpy.float64) for Qdrant compatibility
+                sparse_dict = {k: float(v / norm) for k, v in sparse_dict.items()}
 
         return sparse_dict
